LoginController내 코드를 UserController로 이관 외 관련 코드 변경
