+++ src/main/resources/templates/base.html
@@ -75,7 +75,7 @@
             <ul class="nav navbar-nav navbar-right">
                 <li class="active"><a href="/">Posts</a></li>
                 {{^ loginedUser }}
-                <li><a href="/login/form" role="button">로그인</a></li>
+                <li><a href="/login" role="button">로그인</a></li>
                 <li><a href="/users/form" role="button">회원가입</a></li>
                 {{/ loginedUser }}
                 {{# loginedUser }}
